Fixed compilation for efi app.

+++ include/stddef.h
@@ -3,7 +3,7 @@
 
 #define NULL ((void*)0)
 
-typedef int wchar_t;
+typedef short unsigned int wchar_t;
 typedef long unsigned int size_t;
 typedef long int ssize_t;
 typedef long unsigned int uintptr_t;
